add user & some modifications

+++ src/product/category/category.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CategoryRepository } from './category.repository';
 import { CreateCategoryDto } from './dto/create-category.dto';
 import { UpdateCategoryDto } from './dto/update-category.dto';
@@ -12,7 +12,11 @@ export class CategoryService {
   }
 
   async findById(id: string) {
-    return this.categoryRepository.findById(id);
+    const category = await this.categoryRepository.findById(id);
+    if (!category) {
+      throw new NotFoundException(`Category #${id} not found`);
+    }
+    return category;
   }
 
   async findByName(name: string) {
@@ -24,10 +28,18 @@ export class CategoryService {
   }
 
   async update(id: string, category: UpdateCategoryDto) {
+    const categoryExists = await this.categoryRepository.findById(id);
+    if (!categoryExists) {
+      throw new NotFoundException(`Category #${id} not found`);
+    }
     return this.categoryRepository.update(id, category);
   }
 
   async delete(id: string) {
+    const category = await this.categoryRepository.findById(id);
+    if (!category) {
+      throw new NotFoundException(`Category #${id} not found`);
+    }
     return this.categoryRepository.delete(id);
   }
 }
